[Identity Provider service] [Catalog service] [Photo service] [YARP service]: dependency injection refactoring
