+++ src/BuildingBlocks/BuildingBlocks/Extensions/DistributedCacheExtensions.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace BuildingBlocks.Extensions
+{
+	public static class DistributedCacheExtensions
+	{
+		public static void AddDistributedCache(this IServiceCollection services, IConfiguration configuration)
+		{
+			services.AddStackExchangeRedisCache(options =>
+			{
+				options.Configuration = configuration.GetConnectionString("Redis");
+			});
+		}
+	}
+}
