+++ src/BuildingBlocks/BuildingBlocks/Extensions/ExceptionHandlerExtensions.cs
@@ -1,18 +1,13 @@
 ﻿using BuildingBlocks.Exceptions.Handler;
+using Microsoft.Extensions.DependencyInjection;
 
-namespace Photo.API.Extensions
+namespace BuildingBlocks.Extensions
 {
 	public static class ExceptionHandlerExtensions
 	{
 		public static void AddExceptionHandlerServices(this IServiceCollection services)
 		{
 			services.AddExceptionHandler<CustomExceptionHandler>();
-		}
-
-		public static WebApplication UseCustomExceptionHandler(this WebApplication app)
-		{
-			app.UseExceptionHandler(options => { /* logging etc */ });
-			return app;
 		}
 	}
 }
