+++ src/BuildingBlocks/BuildingBlocks/Extensions/SwaggerExtensions.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.OpenApi.Models;
+
+namespace BuildingBlocks.Extensions
+{
+	public static class SwaggerExtensions
+	{
+		public static void AddSwaggerDocumentation(this IServiceCollection services, string title)
+		{
+			services.AddEndpointsApiExplorer();
+
+			services.AddSwaggerGen(c =>
+			{
+				c.SwaggerDoc("v1", new OpenApiInfo { Title = title, Version = "v1" });
+
+				var jwtSecurityScheme = new OpenApiSecurityScheme
+				{
+					Name = "Authorization",
+					Description = "Enter JWT Bearer token",
+					In = ParameterLocation.Header,
+					Type = SecuritySchemeType.Http,
+					Scheme = "bearer",
+					BearerFormat = "JWT",
+					Reference = new OpenApiReference
+					{
+						Type = ReferenceType.SecurityScheme,
+						Id = "Bearer"
+					}
+				};
+
+				c.AddSecurityDefinition("Bearer", jwtSecurityScheme);
+				c.AddSecurityRequirement(new OpenApiSecurityRequirement
+			{
+				{ jwtSecurityScheme, Array.Empty<string>() }
+			});
+			});
+		}
+	}
+}
