--- src/Services/Catalog/Catalog.API/Configurations/MongoSettings.cs
@@ -1,9 +1,0 @@
-﻿namespace Catalog.API.Configurations
-{
-	public class MongoSettings
-	{
-		public string ConnectionString { get; set; } = null!;
-
-		public string DatabaseName { get; set; } = null!;
-	}
-}
