--- src/Services/Catalog/Catalog.API/Extensions/ExceptionHandlerExtensions.cs
@@ -1,18 +1,0 @@
-﻿using BuildingBlocks.Exceptions.Handler;
-
-namespace Catalog.API.Extensions
-{
-	public static class ExceptionHandlerExtensions
-	{
-		public static void AddExceptionHandlerServices(this IServiceCollection services)
-		{
-			services.AddExceptionHandler<CustomExceptionHandler>();
-		}
-
-		public static WebApplication UseCustomExceptionHandler(this WebApplication app)
-		{
-			app.UseExceptionHandler(options => { /* logging etc */ });
-			return app;
-		}
-	}
-}
