+++ src/Services/Catalog/Catalog.API/Program.cs
@@ -3,26 +3,9 @@ using Catalog.API.Extensions;
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.RegisterApplicationServices(builder.Configuration);
-builder.Services.AddControllers();
 
 var app = builder.Build();
 
-app.UseCustomMiddlewares();
-
-app.MapControllers();
-
-app.UseCors("AllowFromGateway");
-
-if (app.Environment.IsDevelopment())
-{
-	app.UseSwagger();
-	app.UseSwaggerUI(options =>
-	{
-		options.SwaggerEndpoint("/swagger/v1/swagger.json", "Catalog API v1");
-		options.RoutePrefix = string.Empty;
-	});
-
-	await app.InitializeDatabaseAsync();
-}
+await app.UseCustomMiddlewares();
 
 app.Run();
