--- src/Services/IdentityProvider/IdentityProvider.API/Extensions/ExceptionHandlerExtensions.cs
@@ -1,18 +1,0 @@
-﻿using BuildingBlocks.Exceptions.Handler;
-
-namespace IdentityProvider.API.Extensions
-{
-	public static class ExceptionHandlerExtensions
-	{
-		public static void AddExceptionHandlerServices(this IServiceCollection services)
-		{
-			services.AddExceptionHandler<CustomExceptionHandler>();
-		}
-
-		public static WebApplication UseCustomExceptionHandler(this WebApplication app)
-		{
-			app.UseExceptionHandler(options => { /* optional error handling */ });
-			return app;
-		}
-	}
-}
