+++ src/Services/IdentityProvider/IdentityProvider.API/Extensions/ServiceCollectionExtensions.cs
@@ -1,15 +1,18 @@
-﻿namespace IdentityProvider.API.Extensions
+﻿using BuildingBlocks.Extensions;
+
+namespace IdentityProvider.API.Extensions
 {
 	public static class ServiceCollectionExtensions
 	{
 		public static void RegisterApplicationServices(this IServiceCollection services, IConfiguration configuration)
 		{
-			services.AddCorsPolicy();
 			services.AddJwtAuthentication(configuration);
+			services.AddCorsPolicy();
+			services.AddExceptionHandlerServices();
+			services.AddSwaggerDocumentation("Identity Provider API");
 			services.AddIdentityCoreServices(configuration);
 			services.AddDatabaseInfrastructure(configuration);
-			services.AddExceptionHandlerServices();
-			services.AddSwaggerDocumentation();
+			services.AddControllers();
 		}
 	}
 }
