--- src/Services/Photo/Photo.API/Extensions/AuthenticationExtensions.cs
@@ -1,34 +1,0 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using Photo.API.Configurations;
-using System.Text;
-
-namespace Photo.API.Extensions
-{
-	public static class AuthenticationExtensions
-	{
-		public static void AddJwtAuthentication(this IServiceCollection services, IConfiguration configuration)
-		{
-			services.Configure<JwtSettings>(configuration.GetSection("JwtSettings"));
-
-			services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-				.AddJwtBearer(options =>
-				{
-					var jwtSettings = configuration.GetSection("JwtSettings").Get<JwtSettings>();
-					options.RequireHttpsMetadata = false;
-					options.SaveToken = true;
-					options.TokenValidationParameters = new TokenValidationParameters
-					{
-						ValidateIssuer = true,
-						ValidateAudience = true,
-						ValidIssuer = jwtSettings.Issuer,
-						ValidAudience = jwtSettings.Audience,
-						IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtSettings.Secret)),
-						ValidateIssuerSigningKey = true
-					};
-				});
-
-			services.AddAuthorization();
-		}
-	}
-}
