--- src/Services/Photo/Photo.API/Extensions/CorsExtensions.cs
@@ -1,24 +1,0 @@
-﻿namespace Photo.API.Extensions
-{
-	public static class CorsExtensions
-	{
-		public static void AddCorsPolicy(this IServiceCollection services)
-		{
-			services.AddCors(options =>
-			{
-				options.AddDefaultPolicy(builder =>
-					builder.AllowAnyOrigin()
-						   .AllowAnyMethod()
-						   .AllowAnyHeader());
-
-				options.AddPolicy("AllowFromGateway", policy =>
-				{
-					policy.WithOrigins("https://localhost:6061")
-						  .AllowAnyMethod()
-						  .AllowAnyHeader()
-						  .AllowCredentials();
-				});
-			});
-		}
-	}
-}
